Move all components into .jsx files of their won
