+++ src/Components/CastIntroductionVideoLink/index.js
@@ -1,19 +1,3 @@
-import React from "react";
-import { FormattedMessage } from "react-intl";
-import PropTypes from "prop-types";
-import { ContainerLink } from "./styled";
-
-const CastIntroductionVideoLink = ({ href }) => {
-    return (
-        <ContainerLink href={href} target="_blank">
-            [<i className="material-icons">play_circle_outline</i>&nbsp;
-            <FormattedMessage id="cast.introductionVideo" />]
-        </ContainerLink>
-    );
-};
-
-CastIntroductionVideoLink.propTypes = {
-    href: PropTypes.string.isRequired,
-};
+import CastIntroductionVideoLink from './CastIntroductionVideoLink';
 
 export default CastIntroductionVideoLink;
