+++ src/Pages/Rewards/Rewards.jsx
@@ -1,0 +1,13 @@
+import React, { Component } from "react";
+
+class Rewards extends Component {
+    componentDidMount() {
+        window.location.href = "https://www.youtube.com/watch?v=dQw4w9WgXcQ";
+    }
+
+    render() {
+        return <div />;
+    }
+}
+
+export default Rewards;
